requirements.txt
djangorestframework
django
gunicorn
django-heroku
requests